[gardening] Fix MSVC build after a52f2b96179e.
Without this, compiling with MSVC fails with the error:
```
error C2397: conversion from 'int' to 'uint8_t' requires a narrowing conversion
```

TEST=vm-msvc-windows builds

Cq-Include-Trybots: luci.dart.try:vm-msvc-windows-try
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/312262
Auto-Submit: Tess Strickland <sstrickl@google.com>
Commit-Queue: Tess Strickland <sstrickl@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>

+++ runtime/vm/object.h
@@ -3087,7 +3087,7 @@ class Function : public Object {
     uint8_t depth = UntaggedClosureData::kNoAwaiterLinkDepth;
     // Context index at which the `@pragma('vm:awaiter-link')` variable
     // is located.
-    uint8_t index = -1;
+    uint8_t index = static_cast<uint8_t>(-1);
   };
 
   AwaiterLink awaiter_link() const;
